# test-results/2025-09-03/crawler-tests/data/openapi_schema_170814.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "历史文本项目 - 数据源服务",
    "description": "\n    ## 数据源服务 API\n    \n    提供多平台内容爬取、代理管理和内容处理功能的微服务。\n    \n    ### 主要功能\n    - 🕷️ **爬虫管理**: 支持今日头条、百家号、小红书等平台\n    - 📄 **内容管理**: 手动添加、批量导入、查询和更新内容\n    - 🌐 **代理管理**: 代理池管理、自动测试和轮换\n    - 📊 **监控统计**: 实时状态监控和数据统计\n    \n    ### 技术栈\n    - **框架**: FastAPI + Python 3.9+\n    - **数据库**: MongoDB + Redis\n    - **爬虫**: Scrapy + Selenium\n    - **代理**: 多供应商代理池\n    \n    ### 使用说明\n    1. 使用 `/crawlers/` 接口管理爬虫任务\n    2. 使用 `/content/` 接口管理内容数据\n    3. 使用 `/proxy/` 接口管理代理设置\n    4. 查看 `/health` 检查服务状态\n    ",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/crawlers/": {
      "post": {
        "tags": [
          "爬虫管理"
        ],
        "summary": "创建爬虫任务",
        "description": "创建新的爬虫任务\n\n- **platform**: 平台类型 (toutiao, baijiahao, xiaohongshu)\n- **keywords**: 关键词列表，用于内容抓取\n- **max_pages**: 最大抓取页数\n- **interval**: 请求间隔时间，避免过于频繁请求\n- **priority**: 任务优先级，影响执行顺序",
        "operationId": "create_crawler_task_api_v1_crawlers__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCrawlerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Crawler Task Api V1 Crawlers  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "爬虫管理"
        ],
        "summary": "获取所有任务",
        "description": "获取所有爬虫任务列表，支持分页和过滤",
        "operationId": "get_all_crawler_tasks_api_v1_crawlers__get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CrawlerStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按状态过滤",
              "title": "Status"
            },
            "description": "按状态过滤"
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ContentSource"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按平台过滤",
              "title": "Platform"
            },
            "description": "按平台过滤"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码",
              "default": 1,
              "title": "Page"
            },
            "description": "页码"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页数量",
              "default": 20,
              "title": "Size"
            },
            "description": "每页数量"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get All Crawler Tasks Api V1 Crawlers  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crawlers/{task_id}/start": {
      "post": {
        "tags": [
          "爬虫管理"
        ],
        "summary": "启动爬虫任务",
        "description": "启动指定的爬虫任务\n\n- **task_id**: 爬虫任务的唯一标识符",
        "operationId": "start_crawler_task_api_v1_crawlers__task_id__start_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "任务ID",
              "title": "Task Id"
            },
            "description": "任务ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Start Crawler Task Api V1 Crawlers  Task Id  Start Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crawlers/{task_id}/stop": {
      "post": {
        "tags": [
          "爬虫管理"
        ],
        "summary": "停止爬虫任务",
        "description": "停止指定的爬虫任务",
        "operationId": "stop_crawler_task_api_v1_crawlers__task_id__stop_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "任务ID",
              "title": "Task Id"
            },
            "description": "任务ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Stop Crawler Task Api V1 Crawlers  Task Id  Stop Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crawlers/{task_id}/pause": {
      "post": {
        "tags": [
          "爬虫管理"
        ],
        "summary": "暂停爬虫任务",
        "description": "暂停指定的爬虫任务",
        "operationId": "pause_crawler_task_api_v1_crawlers__task_id__pause_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "任务ID",
              "title": "Task Id"
            },
            "description": "任务ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Pause Crawler Task Api V1 Crawlers  Task Id  Pause Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crawlers/{task_id}/resume": {
      "post": {
        "tags": [
          "爬虫管理"
        ],
        "summary": "恢复爬虫任务",
        "description": "恢复暂停的爬虫任务",
        "operationId": "resume_crawler_task_api_v1_crawlers__task_id__resume_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "任务ID",
              "title": "Task Id"
            },
            "description": "任务ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Resume Crawler Task Api V1 Crawlers  Task Id  Resume Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crawlers/{task_id}/status": {
      "get": {
        "tags": [
          "爬虫管理"
        ],
        "summary": "获取任务状态",
        "description": "获取爬虫任务的详细状态信息",
        "operationId": "get_crawler_task_status_api_v1_crawlers__task_id__status_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "任务ID",
              "title": "Task Id"
            },
            "description": "任务ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Crawler Task Status Api V1 Crawlers  Task Id  Status Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/crawlers/{task_id}": {
      "delete": {
        "tags": [
          "爬虫管理"
        ],
        "summary": "删除任务",
        "description": "删除指定的爬虫任务（仅限已完成或已停止的任务）",
        "operationId": "delete_crawler_task_api_v1_crawlers__task_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "任务ID",
              "title": "Task Id"
            },
            "description": "任务ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Delete Crawler Task Api V1 Crawlers  Task Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContentSource": {
        "type": "string",
        "enum": [
          "toutiao",
          "baijiahao",
          "xiaohongshu",
          "manual",
          "rss",
          "api"
        ],
        "title": "ContentSource",
        "description": "内容来源枚举"
      },
...
      "CrawlerStatus": {
        "type": "string",
        "enum": [
          "idle",
          "running",
          "paused",
          "stopped",
          "error",
          "finished"
        ],
        "title": "CrawlerStatus",
        "description": "爬虫状态枚举"
      },
      "CreateCrawlerRequest": {
        "properties": {
          "platform": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ContentSource"
              }
            ],
            "description": "平台类型"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Keywords",
            "description": "关键词列表"
          },
          "max_pages": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Pages",
            "description": "最大抓取页数",
            "default": 10
          },
          "interval": {
            "type": "number",
            "maximum": 60.0,
            "minimum": 0.1,
            "title": "Interval",
            "description": "请求间隔(秒)",
            "default": 5.0
          },
          "timeout": {
            "type": "integer",
            "maximum": 3600.0,
            "minimum": 30.0,
            "title": "Timeout",
            "description": "超时时间(秒)",
            "default": 300
          },
          "priority": {
            "allOf": [
              {
                "$ref": "#/components/schemas/TaskPriority"
              }
            ],
            "description": "任务优先级",
            "default": 5
          },
          "proxy_enabled": {
            "type": "boolean",
            "title": "Proxy Enabled",
            "description": "是否启用代理",
            "default": true
          },
          "retry_attempts": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Retry Attempts",
            "description": "重试次数",
            "default": 3
          },
          "custom_settings": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Settings",
            "description": "自定义配置"
          }
        },
        "type": "object",
        "required": [
          "platform",
          "keywords"
        ],
        "title": "CreateCrawlerRequest",
        "description": "创建爬虫任务请求模型",
        "example": {
          "custom_settings": {},
          "interval": 5.0,
          "keywords": [
            "历史",
            "文化"
          ],
          "max_pages": 10,
          "platform": "toutiao",
          "priority": 5,
          "proxy_enabled": true,
          "retry_attempts": 3,
          "timeout": 300
        }
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskPriority": {
        "type": "integer",
        "enum": [
          1,
          5,
          8,
          10
        ],
        "title": "TaskPriority",
        "description": "任务优先级枚举"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}